using custom hook to manage selected actions
