+++ src/components/groups/custom/GroupMembers.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React, { FormEvent, useState } from "react";
+import React, { useState } from "react";
 import InnerContainer from "../../utils/InnerContainer";
 import GroupMemberTab from "./GroupMemberTab";
 import {
@@ -15,8 +15,6 @@ import {
 } from "@/src/utils/constants";
 import InviteMemberForm from "./InviteMemberForm";
 import HRow from "../../utils/HRow";
-import { toast } from "sonner";
-import { useRouter } from "next/navigation";
 import { LoadingIconBright } from "../../utils/LoadingIcon";
 import SecondaryButton from "../../utils/SecondaryButton";
 import AnnouncementTag from "../../announcements/AnnouncementTag";
@@ -24,18 +22,7 @@ import FormInputContainer from "../../utils/FormInputContainer";
 import PrimaryButton from "../../utils/PrimaryButton";
 import Modal from "../../utils/Modal";
 import ModalLoading from "../../utils/ModalLoading";
-import { handleModify } from "@/src/hooks/groups/custom/handleSelectActions";
-import {
-  manageAdmin,
-  removeMembers,
-} from "@/src/hooks/groups/custom/handleMembers";
-
-const DEFAULT_SELECT = {
-  state: false,
-  members: [] as string[],
-};
-
-const DEFAULT_ACTION = Object.keys(GROUP_MEMBERS_SELECT_OPTIONS)[0];
+import { useHandleSelectActions } from "@/src/hooks/groups/custom/useHandleSelectActions";
 
 export default function GroupMembers({
   membersList,
@@ -46,7 +33,6 @@ export default function GroupMembers({
   groupID: string;
   curMember: GROUP_MEMBERS_SCHEMA;
 }) {
-  const router = useRouter();
   const [show, setShow] = useState(false);
   const [online, setOnline] = useState(0);
   const { itemList, search, handleSearch } = useQueryObj({ obj: membersList });
@@ -58,172 +44,10 @@ export default function GroupMembers({
     (id: string) => membersList[id].role === "owner"
   )[0];
 
-  const [loading, setLoading] = useState(false);
-
-  const [actions, setActions] = useState(DEFAULT_ACTION);
-
-  const [select, setSelect] = useState(DEFAULT_SELECT);
-
-  const reset = () => {
-    setSelect(DEFAULT_SELECT);
-    setActions(DEFAULT_ACTION);
-  };
-
-  const confirmBookIn = async () => {
-    if (
-      confirm(
-        "Are you sure you want to book in selected members of this group?"
-      )
-    ) {
-      setLoading(true);
-      try {
-        const res = await handleModify(select.members, { bookedIn: true });
-        if (!res.status) throw new Error(res.error);
-        toast.success("Successfully welcomed back selected members.");
-        router.refresh();
-      } catch (err: any) {
-        toast.error(err.message);
-      }
-      setLoading(false);
-    }
-  };
-
-  const confirmBookOut = async () => {
-    if (
-      confirm(
-        "Are you sure you want to book out selected members of this group?"
-      )
-    ) {
-      setLoading(true);
-      try {
-        const res = await handleModify(select.members, { bookedIn: false });
-        if (!res.status) throw new Error(res.error);
-        toast.success("Happy book out selected members!");
-        router.refresh();
-      } catch (err: any) {
-        toast.error(err.message);
-      }
-      setLoading(false);
-    }
-  };
-
-  const confirmCourse = async () => {
-    if (
-      confirm("Are you sure you want to mark the selected members on-course?")
-    ) {
-      setLoading(true);
-      try {
-        const res = await handleModify(select.members, {
-          isOnCourse: true,
-          bookedIn: false,
-        });
-        if (!res.status) throw new Error(res.error);
-        toast.success("Great, we marked these members as on-course.");
-        router.refresh();
-      } catch (err: any) {
-        toast.error(err.message);
-      }
-      setLoading(false);
-    }
-  };
-
-  const confirmUncourse = async () => {
-    if (
-      confirm("Are you sure you want to unmark the selected members on-course?")
-    ) {
-      setLoading(true);
-      try {
-        const res = await handleModify(select.members, {
-          isOnCourse: false,
-          bookedIn: true,
-        });
-        if (!res.status) throw new Error(res.error);
-        toast.success("Welcome back, members as on-course!");
-        router.refresh();
-        setSelect({ members: [], state: false });
-      } catch (err: any) {
-        toast.error(err.message);
-      }
-      setLoading(false);
-    }
-  };
-
-  const confirmRemove = async () => {
-    if (
-      confirm("Are you sure you want to remove these members from this group?")
-    ) {
-      setLoading(true);
-      try {
-        const res = await removeMembers(groupID, select.members);
-        if (!res.status) throw new Error(res.error);
-        toast.success("Goodbye to those members.");
-        router.refresh();
-      } catch (err: any) {
-        toast.error(err.message);
-      }
-      setLoading(false);
-    }
-  };
-
-  const confirmAdmin = async () => {
-    if (confirm("Are you sure you want to make these members admins?")) {
-      setLoading(true);
-      try {
-        const res = await manageAdmin(groupID, select.members, true);
-        if (!res.status) throw new Error(res.error);
-        toast.success("Great, new admins here!");
-        router.refresh();
-      } catch (err: any) {
-        toast.error(err.message);
-      }
-      setLoading(false);
-    }
-  };
-
-  const confirmRemAdmin = async () => {
-    if (
-      confirm("Are you sure you want to remove permissions from these members?")
-    ) {
-      setLoading(true);
-      try {
-        const res = await manageAdmin(groupID, select.members, false);
-        if (!res.status) throw new Error(res.error);
-        toast.success("Well, that was unfortunate for those members.");
-        router.refresh();
-      } catch (err: any) {
-        toast.error(err.message);
-      }
-      setLoading(false);
-    }
-  };
-
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-    try {
-      if (actions === "Book In") {
-        await confirmBookIn();
-      } else if (actions === "Book Out") {
-        await confirmBookOut();
-      } else if (actions === "Mark as On-Course") {
-        await confirmCourse();
-      } else if (actions === "Unmark as On-Course") {
-        await confirmUncourse();
-      } else if (actions === "Remove from Group") {
-        await confirmRemove();
-      } else if (actions === "Make Admin") {
-        await confirmAdmin();
-      } else if (actions === "Remove Admin") {
-        await confirmRemAdmin();
-      }
-    } catch (err: any) {
-      toast.error(err.message);
-    }
-    setLoading(false);
-    reset();
-  };
-
   const addOnline = () => setOnline((online) => online + 1);
+
+  const { loading, handleSubmit, select, setActions, setSelect } =
+    useHandleSelectActions(groupID);
 
   return (
     <>
@@ -249,10 +73,6 @@ export default function GroupMembers({
               handleSearch={handleSearch}
               search={search}
             />
-            {/* <div className="flex items-center justify-end gap-1 w-fit">
-            <p className="text-xs text-custom-grey-text">{online} in camp</p>
-            <StatusDot className="w-1 h-1" status />
-          </div> */}
             <InnerContainer className="max-h-[60vh] relative">
               {isAdmin && (
                 <div className="w-full flex flex-col gap-2 items-start justify-start p-2 pt-0 sticky top-0 left-0 z-10 bg-white shadow-sm">
